requirements.txt
Django>=5.2.0,<5.3.0
djangorestframework>=3.14.0,<3.15.0
django-cors-headers>=4.3.0,<4.4.0
python-dotenv>=1.0.0,<1.1.0
gunicorn>=21.2.0,<21.3.0
